Add physics/jump mechanics, change viewport again, run RPi code in background
The Lisse node is now using `_physics_process` instead of `_process`, and now has
features such as gravity, friction and can now double jump. A floor has been added
so that Lisse doesn't fall to their death.

The previous approach with drawing portrait graphics didn't work when working with
`_physics_process`, as those seem to be applied after the rotation has been done.
Instead everything is now rendered to a rotated `TextureRect` using the output of a
`GameViewPort` which contains the entire project. Additionally the window now has a
manual override to match the rotated output.

The Raspberry Pi code has been moved to a global node to ensure it always runs. Also
it's nice to not need to have it as a node.