Read screen edges to LEDs
This commit adds support for reading and sending the edges of
the screen to the LED strip. This works by extending the view-
port width by two pixels, and then sampling the new columns at
certain intervals.

The idea is to be able to animate the LEDs by animating things
on these extra columns, which should be a lot easier than hard-
coding all colors. This also works as an AmbiLight from Wish if
the scene is extended to draw in these two columns.